qlib-agent-framework/requirements.txt
pyqlib
numpy
cython>=0.29
lightgbm
torch
openai
pyyaml
pandas
tqdm
